feat：优化第三方构建机调度策略，新增标签调度 #7342

+++ src/backend/ci/core/environment/api-environment/src/main/kotlin/com/tencent/devops/environment/api/ServiceLabelResource.kt
@@ -60,5 +60,5 @@ interface ServiceLabelResource {
         projectId: String,
         @ApiParam(value = "计算表达式列表", required = true)
         calculateExpression: CalculateExpression
-    ): Result<List<Int>>
+    ): Result<List<Long>>
 }
